Cleaned up differences between io and mqtt classes, fixed up metrics on browser
